# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  shareds.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  status.StatusRequest:
    properties:
      creation_date:
        description: 'Data de criação (padrão: data atual)'
        type: string
      modification_date:
        description: Data de modificação (opcional)
        type: string
      name:
        description: Nome do status (único)
        type: string
      status_uuid:
        description: UUID do status (chave primária)
        type: string
    type: object
  status.StatusResponse:
    properties:
      creation_date:
        description: 'Data de criação (padrão: data atual)'
        type: string
      modification_date:
        description: Data de modificação (opcional)
        type: string
      name:
        description: Nome do status (único)
        type: string
      status_uuid:
        description: UUID do status (chave primária)
        type: string
    type: object
  users.UserRequest:
    description: Struct containing the data required to create/update a user
    properties:
      email:
        example: test@test.com
        type: string
      password:
        example: Password123#
        type: string
      phone:
        example: ""
        type: string
      position:
        example: ""
        type: string
      position_uuid:
        example: ""
        type: string
      status_uuid:
        example: ""
        type: string
      tax_number:
        example: "12345678901"
        type: string
      user_uuid:
        example: user_uuid
        type: string
      username:
        example: username
        type: string
    required:
    - email
    - username
    type: object
  users.UserResponse:
    properties:
      ModificationDate:
        type: string
      creation_date:
        type: string
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      position:
        type: string
      profile_image_link:
        type: string
      status_uuid:
        type: string
      tax_number:
        type: string
      user_uuid:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: Bernardtm Backend
  title: backend
  version: "1.0"
paths:
...
  /status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/status.StatusResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all Status
      tags:
      - Status
    post:
      consumes:
      - application/json
      parameters:
      - description: Status Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/status.StatusRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new Status
      tags:
      - Status
  /status/{id}:
    delete:
      parameters:
      - description: ID of the Status
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Status by ID
      tags:
      - Status
    get:
      parameters:
      - description: ID of the Status
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/status.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Status by ID
      tags:
      - Status
    put:
      consumes:
      - application/json
      parameters:
      - description: ID of the Status
        in: path
        name: id
        required: true
        type: string
      - description: Updated Status Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/status.StatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an existing Status
      tags:
      - Status
  /status/paginate:
    get:
      parameters:
      - description: Page Number
        in: query
        name: page
        required: true
        type: integer
      - description: Page Size
        in: query
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/status.StatusResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Paginate Status
      tags:
      - Status
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/users.UserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all Users
      tags:
      - Users
    post:
      consumes:
      - application/json
      parameters:
      - description: User Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/users.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new User
      tags:
      - Users
  /users/{id}:
    delete:
      parameters:
      - description: ID of the User
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete User by ID
      tags:
      - Users
    get:
      parameters:
      - description: ID of the User
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get User by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      parameters:
      - description: ID of the User
        in: path
        name: id
        required: true
        type: string
      - description: Updated User Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/users.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an existing User
      tags:
      - Users
  /users/paginate:
    get:
      parameters:
      - description: Page Number
        in: query
        name: page
        required: true
        type: integer
      - description: Page Size
        in: query
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/users.UserResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shareds.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Paginate Users
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: 'Provide the JWT token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"